% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function5_alterdesc.R
\name{alter_desc_pnd}
\alias{alter_desc_pnd}
\title{Compositional descriptives of persnet data object}
\usage{
alter_desc_pnd(persnet_df)
}
\arguments{
\item{persnet_list_igra}{}
}
\description{
Computes percentage compositions of various alter characteristics in a persnet data object
}
