\name{organize_pnd}
\alias{organize_pnd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
organize pnd
}
\description{
This function takes in a dataframe of persnet responses and returns a list of igraph objects.
}
\usage{
organize_pnd(persnet_df)
}
\arguments{
  \item{persnet_df}{
}
}
\details{
}
\value{
}
\references{
}

}
\note{
}
\seealso{
}
\examples{
}

