% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function4_netdesc.R
\name{net_desc_pnd}
\alias{net_desc_pnd}
\title{Network descriptives of persnet data object}
\usage{
net_desc_pnd(persnet_dataframe)
}
\arguments{
\item{persnet_dataframe}{}
}
\description{
Computes network statistics of persnet data object
}
