\name{organize_pnd}
\alias{organize_pnd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
organize pnd
}
\description{
This function takes in a dataframe of persnet responses and returns a list of igraph objects. Within each igraph object, the nodes have the attributes of the name specified in the survey, an indicator variable defining whether they are an alter or not, and each edge is a weighted edge.
}
\usage{
organize_pnd(persnet_df)
}
\arguments{
  \item{persnet_df}{
}
}
\details{
}
\value{
}
\references{
}

}
\note{
}
\seealso{
}
\examples{
}

